% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{A Simple Hello Function}
\usage{
hello(name)
}
\arguments{
\item{name}{A character string representing the user's name.}
}
\value{
A character string greeting the user.
}
\description{
This function greets the user.
}
\examples{
hello("Alice")
}
